Move DailyRepresentativeValues to own submodule, updated methods
