+++ pyrealm/subdaily/__init__.py
@@ -1,0 +1,9 @@
+"""The :mod:`~pyrealm.subdaily` module provides extensions of the P Model that allow for
+the explicit modelling of the fast and slow responses of photosynthesis to changing
+conditions.
+
+TODO.
+"""  # noqa: D205, D415
+
+# Flatten submodule namespace
+from pyrealm.subdaily.daily_values import DailyRepresentativeValues  # noqa: F401
